Add files via upload
The code is not refined or polished in any way, it has the necessary components to run and a couple of extra features such as the "blackjack" and "bust" writings, though they are still there as I was experimenting with color mechanics. The method used to create the whole UI is solely PyGame. The game actually wholly runs on the console as well, as I first built its entirety without any concern about gameplay. Some texts, for example, go out of boundaries and the image qualities nor the controls pass the mark of being "enough". Since these aspects were and are still not the main objective, this version of the code is satisfactory.